Move auth test helper to proper folder
